Bgk 63 fix

+++ BGK.Business/Models/CompanyModel.cs
@@ -1,9 +1,4 @@
 ﻿using BearGoodbyeKolkhozProject.Data.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BearGoodbyeKolkhozProject.Business.Models
 {
@@ -11,13 +6,13 @@ namespace BearGoodbyeKolkhozProject.Business.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Tin { get; set; }
+        public long Tin { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public string Password { get; set; }
         public bool IsDeleted { get; set; }
 
-        public  ICollection<TrainingReview> TrainingReviews { get; set; }
-        public  ICollection<LecturerReview> LecturerReviews { get; set; }
+        public ICollection<TrainingReview> TrainingReviews { get; set; }
+        public ICollection<LecturerReview> LecturerReviews { get; set; }
     }
 }
